fix(go): code gen not properly using FQN type
Fixes broken type signature in class creation to use `FQN` type instead
of `string` which was previously used for `interfaces` list.

Refactor jsii-pacmak go `runtime` module to separate definitions out of
`index.ts`.
